Conflicts:
	app/assets/javascripts/game.js

Merge Ed code with Alberto

+++ app/assets/javascripts/game.js
@@ -1,5 +1,11 @@
 'use strict';
 
+//overriding the framework so rick doesn't get stuck on those platform transitions
+Phaser.Physics.Arcade.prototype.separate = function (body1, body2) {
+
+        this._result = (this.separateY(body1, body2) || this.separateX(body1, body2));
+    }
+
 Rick.Game = function (game) {
 
   // When a State is added to Phaser it automatically has the following properties set on it, even if they already exist:
@@ -45,7 +51,7 @@ Rick.Game = function (game) {
   this.nextEnemyTime = 3000; // time span. Will decrease to increase difficult level
   this.enemyKillPoint = 20;
 
-  // levels
+  // levels (of difficulty)
   this.levelTime;
   this.changeLevelTime = 5000; // interval
 
@@ -60,6 +66,10 @@ Rick.Game = function (game) {
   this.score = 0;
   this.scoreString;
   this.scoreText;
+
+
+  //Lives
+  this.lives;
 
 
   //	You can use any of these from any function within this State.
@@ -101,19 +111,7 @@ Rick.Game.prototype = {
     this.platforms.add(this.platform);
 
     // Create player
-    this.player = this.game.add.sprite(100, 0, 'rick');
-    this.player.body.setSize(60, 90, 0, 0);
-    this.player.anchor.setTo(0.5, 0.5);
-
-    // this.player.body.bounce.y = 0.3;
-
-    // this.player.body.bounce.y = 0.3;
-    this.player.body.gravity.y = 10;
-    // set collideWorldBounds to left only, or kill player on touching bottom
-    this.player.body.collideWorldBounds = false;
-
-    this.player.animations.add('right', [0,1,2,3,4,5,6,7], 10, true);
-    this.player.animations.add('jump', [8], 10, false);
+    this.createPlayer();
 
     // Adds Keyboard controls
     this.keyboard = this.game.input.keyboard.createCursorKeys();
@@ -135,6 +133,18 @@ Rick.Game.prototype = {
     // The score
     this.scoreString = 'Score : ';
     this.scoreText = this.game.add.text(10, 10, this.scoreString + this.score, { fontSize: '34px', fill: '#fff' });
+
+    // Lives display
+    this.lives = this.game.add.group();
+    this.game.add.text(10, 450, 'Lives : ', { fontSize: '34px', fill: '#fff' });
+
+    // The 3 lives as objects (head)
+    for (var i = 0; i < 3; i++) 
+    {
+        var head = this.lives.create(120 + (50 * i), 460, 'head');
+        head.anchor.setTo(0.5, 0.5);
+        head.alpha = 0.4;
+    }
 
 	  // Add Player Statistics
     this.playerStats($('.score_div'));
@@ -207,8 +217,25 @@ Rick.Game.prototype = {
   },
 
   collisionHandlerHitEnemy: function(player, enemy) {
-  	player.kill();
   	enemy.kill();
+
+  	// get the first head (out of the 3 that exist)
+  	var live = this.lives.getFirstAlive();
+
+  	// if any lives exist, kill them
+  	if (live)
+    {
+        live.kill();
+        player.kill();
+        this.createPlayer();
+
+    }
+
+  	// When the player dies
+    if (this.lives.countLiving() < 1){
+    	player.kill();
+    	this.quitGame();
+    }
 
   	//  And create an explosion :)
     var explosion = this.explosions.getFirstDead();
@@ -219,18 +246,41 @@ Rick.Game.prototype = {
     explosion = this.explosions.getFirstDead();
     explosion.reset(player.body.x + 50, player.body.y + 30);
     explosion.play('explosion', 30, false, true);
+
   },
 
-  collisionHandlerFall: function(player){
-  	
 
-  	//  And create an explosion :)
+
+  collisionHandlerFall: function(player){
+
+  	
+    // get the first head (out of the 3 that exist)
+  	var live = this.lives.getFirstAlive();
+
+  	// if any lives exist, kill them
+  	if (live)
+    {
+        live.kill();
+        player.kill();
+        this.createPlayer();
+        // this stops multiple deaths when he falls
+        this.dead = false;
+
+    }
+  	// When the player dies
+    if (this.lives.countLiving() < 1){
+    	player.kill();
+    	this.quitGame();
+    	this.dead = true;
+    	// this stops multiple deaths when he falls, set to false everywhere else
+    }
+
+    //  And create an explosion :)
     var explosion = this.explosions.getFirstDead();
     explosion.reset(player.body.x + 50, player.body.y + 30);
     explosion.play('explosion', 30, false, true);
     
-    player.kill();
-    this.dead = true;
+   
   },
 
   setUpExplosions: function(explosion) {
@@ -246,6 +296,20 @@ Rick.Game.prototype = {
 
     //	Then let's go back to the main menu.
     this.game.state.start('Game');
+    this.dead = false;
+  },
+
+   createPlayer: function () {
+  	// Create player
+    this.player = this.game.add.sprite(100, 0, 'rick');
+    this.player.body.setSize(60, 90, 0, 0);
+    this.player.anchor.setTo(0.5, 0.5);
+    this.player.body.gravity.y = 10;
+    // player will still die, but will survive if in the air at the time
+    this.player.body.collideWorldBounds = true;
+
+    this.player.animations.add('right', [0,1,2,3,4,5,6,7], 10, true);
+    this.player.animations.add('jump', [8], 10, false);
   },
 
   createEnemy: function () {
@@ -262,7 +326,7 @@ Rick.Game.prototype = {
       if (this.enemy) {
         var xPos = [400, 450, 500];
         var yPos = [100, 150, 200, 250];
-        this.enemy.reset(xPos[this.getRandom(0, xPos.length)], yPos[this.getRandom(0, yPos.length)]);
+        this.enemy.reset(xPos[this.getRandom(0, xPos.length - 1)], yPos[this.getRandom(0, yPos.length - 1)]);
         this.enemy.body.velocity.x = -200;
         this.enemiesTime = this.game.time.now + this.nextEnemyTime;
       }
@@ -285,7 +349,7 @@ Rick.Game.prototype = {
         var xPos = [800, 850, 900];
         var yPos = [350, 400, 450];
         this.platform.scale.setTo(2,2);
-        this.platform.reset(xPos[this.getRandom(0, xPos.length)], yPos[this.getRandom(0, yPos.length)]);
+        this.platform.reset(xPos[this.getRandom(0, xPos.length - 1)], yPos[this.getRandom(0, yPos.length - 1)]);
         this.platform.body.velocity.x = this.platformVelocity;
         this.platform.body.immovable = true;
         this.platformsTime = this.game.time.now + 700;
